GEMINI LIVE PROJECT/requirements.txt
google-generativeai
pydantic>=2.0
sentence-transformers
tavily-python
python-dotenv
numpy